+++ packages/hardhat-core/test/internal/hardhat-network/helpers/transactions.ts
@@ -2,7 +2,7 @@ import { Transaction } from "@ethereumjs/tx";
 import { bufferToHex, toBuffer, zeroAddress } from "ethereumjs-util";
 
 import { numberToRpcQuantity } from "../../../../src/internal/core/jsonrpc/types/base-types";
-import { RpcTransactionRequestInput } from "../../../../src/internal/core/jsonrpc/types/input/transactionRequest.s";
+import { RpcTransactionRequestInput } from "../../../../src/internal/core/jsonrpc/types/input/transactionRequest";
 import { TransactionParams } from "../../../../src/internal/hardhat-network/provider/node-types";
 import { HardhatNetworkProvider } from "../../../../src/internal/hardhat-network/provider/provider";
 import { EthereumProvider } from "../../../../src/types";
